[DUX-3046]: Make Slacklinker backlink Linear tickets
Linear is a fine bug tracker and has a Slack integration that is quite
Slacklinker-like, with a feature called "Linkback" which detects ticket IDs and
provides proper links. However, for some reason, that feature doesn't actually
add the link to the mentioning Slack messages on the Linear tickets.

It is kind of important to have proper backlinks to be able to keep track of
ticket context and find out if there are recurring issues.

Thus, we add one.